refactor(cli): rename fury-cli to cli
