--- packages/fury-cli/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "fury-cli",
-  "version": "0.9.4",
-  "description": "Command line tool interface for Fury.js",
-  "author": "Apiary.io <support@apiary.io>",
-  "license": "MIT",
-  "bin": {
-    "fury": "lib/fury.js"
-  },
-  "files": [
-    "lib/fury.js"
-  ],
-  "homepage": "https://github.com/apiaryio/api-elements.js/tree/master/packages/fury-cli",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/apiaryio/api-elements.js.git",
-    "directory": "packages/fury-cli"
-  },
-  "scripts": {
-    "lint": "eslint .",
-    "lint:fix": "eslint . --fix",
-    "test": "exit 0"
-  },
-  "dependencies": {
-    "@apielements/apiaryb-parser": "^0.1.0",
-    "@apielements/apib-parser": "^0.19.0",
-    "@apielements/apib-serializer": "^0.15.0",
-    "@apielements/core": "^0.1.0",
-    "@apielements/openapi2-parser": "^0.31.1",
-    "@apielements/openapi3-parser": "^0.12.3",
-    "cardinal": "^2.1.1",
-    "commander": "^5.1.0",
-    "js-yaml": "^3.12.0",
-    "minim": "^0.23.4"
-  },
-  "devDependencies": {
-    "eslint": "^5.16.0"
-  },
-  "engines": {
-    "node": ">=8"
-  }
-}
